remove stage file
